src/requirements.txt
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
unidecode